Add UpdateStrategy for k8s spec v2 and v3;
