+++ caas/kubernetes/provider/specs/v3_test.go
@@ -166,9 +166,14 @@ configMaps:
     foo: bar
     hello: world
 service:
-  scalePolicy: serial
   annotations:
     foo: bar
+  scalePolicy: serial
+  updateStrategy:
+    type: Recreate
+    rollingUpdate:
+      maxUnavailable: 10%
+      maxSurge: 25%
 serviceAccount:
   automountServiceAccountToken: true
   roles:
@@ -419,8 +424,15 @@ foo: bar
 	getExpectedPodSpecBase := func() *specs.PodSpec {
 		pSpecs := &specs.PodSpec{ServiceAccount: sa1}
 		pSpecs.Service = &specs.ServiceSpec{
-			ScalePolicy: "serial",
 			Annotations: map[string]string{"foo": "bar"},
+			ScalePolicy: "serial",
+			UpdateStrategy: specs.UpdateStrategy{
+				Type: "Recreate",
+				RollingUpdate: &specs.RollingUpdateSpec{
+					MaxUnavailable: &specs.IntOrString{Type: specs.String, StrVal: "10%"},
+					MaxSurge:       &specs.IntOrString{Type: specs.String, StrVal: "25%"},
+				},
+			},
 		}
 		pSpecs.ConfigMaps = map[string]specs.ConfigMap{
 			"mydata": {
